requirements.txt
fastapi
uvicorn
motor
pydantic
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
email-validator
